forward and inversion complete for VMD source.
