+++ tutorials/assets/em1dfm_data.obs
@@ -1,0 +1,5 @@
+3.8200e+02 9.0138e+02 7.2702e+02
+1.8220e+03 1.6540e+03 1.1406e+03
+7.9700e+03 3.0179e+03 1.9620e+03
+3.5920e+04 5.6392e+03 2.0043e+03
+1.3010e+05 6.8207e+03 1.3545e+03
