+++ tutorials/assets/em1dtm_data.obs
@@ -1,0 +1,31 @@
+1.0000e-05 6.0424e-07
+1.2589e-05 5.4259e-07
+1.5849e-05 4.6850e-07
+1.9953e-05 3.9459e-07
+2.5119e-05 3.3877e-07
+3.1623e-05 2.8420e-07
+3.9811e-05 2.3366e-07
+5.0119e-05 1.9199e-07
+6.3096e-05 1.5131e-07
+7.9433e-05 1.2753e-07
+1.0000e-04 1.0592e-07
+1.2589e-04 8.8744e-08
+1.5849e-04 7.6705e-08
+1.9953e-04 6.4267e-08
+2.5119e-04 5.8578e-08
+3.1623e-04 5.1666e-08
+3.9811e-04 4.4620e-08
+5.0119e-04 3.9279e-08
+6.3096e-04 3.4974e-08
+7.9433e-04 2.9832e-08
+1.0000e-03 2.5995e-08
+1.2589e-03 2.1381e-08
+1.5849e-03 1.7777e-08
+1.9953e-03 1.4320e-08
+2.5119e-03 1.0753e-08
+3.1623e-03 8.0587e-09
+3.9811e-03 6.1455e-09
+5.0119e-03 4.3226e-09
+6.3096e-03 3.0812e-09
+7.9433e-03 2.0688e-09
+1.0000e-02 1.4080e-09
