+++ tutorials/plot_1_tdem_fwd_chargeable.py
@@ -1,6 +1,6 @@
 """
-Forward Simulation with Different Waveforms
-===========================================
+Forward Simulation over a Chargeable Earth
+==========================================
 
 
 
@@ -18,7 +18,6 @@ from matplotlib import pyplot as plt
 
 from SimPEG import maps
 import simpegEM1D as em1d
-from simpegEM1D.waveforms import TriangleFun, VTEMFun
 from simpegEM1D.analytics import ColeCole
 
 
@@ -37,84 +36,34 @@ receiver_location = np.array([0., 0., 0.])
 receiver_orientation = "z"  # "x", "y" or "z"
 field_type = "secondary"  # "secondary", "total" or "ppm"
 
-times = np.logspace(-5, -2, 41)
+times = np.logspace(-6, -1, 51)
 
 # Receiver list
-receiver_list = [
+receiver_list = []
+receiver_list.append(
+    em1d.receivers.TimeDomainPointReceiver(
+        receiver_location, times, orientation=receiver_orientation,
+        component="b"
+    )
+)
+receiver_list.append(
     em1d.receivers.TimeDomainPointReceiver(
         receiver_location, times, orientation=receiver_orientation,
         component="dbdt"
     )
-]
+)
 
 # Sources
-source_list = []
-
-# Step off
-source_list.append(
+source_list = [
     em1d.sources.TimeDomainHorizontalLoopSource(
         receiver_list=receiver_list, location=source_location,
-        I=source_current, a=source_radius, wave_type="stepoff"
+        I=source_current, a=source_radius
     )
-)
-
-# Triangle Waveform
-time_input_currents = np.r_[-np.logspace(-2, -5, 31), 0.]
-input_currents = TriangleFun(time_input_currents+0.01, 5e-3, 0.01)
-source_list.append(
-    em1d.sources.TimeDomainHorizontalLoopSource(
-        receiver_list=receiver_list, location=source_location,
-        I=source_current,
-        a=source_radius,
-        wave_type="general",
-        time_input_currents=time_input_currents,
-        input_currents=input_currents,
-        n_pulse = 1,
-        base_frequency = 25.,
-        use_lowpass_filter=False,
-        high_cut_frequency=210*1e3
-    )
-)
-
-# VTEM waveform
-time_input_currents_2 = np.r_[-np.logspace(-2, -5, 31), 0.]
-input_currents_2 = VTEMFun(time_input_currents+0.01, 8e-3, 0.01, 1)
-source_list.append(
-    em1d.sources.TimeDomainHorizontalLoopSource(
-        receiver_list=receiver_list, location=source_location,
-        I=source_current,
-        a=source_radius,
-        wave_type="general",
-        time_input_currents=time_input_currents_2,
-        input_currents=input_currents_2,
-        n_pulse = 1,
-        base_frequency = 25.,
-        use_lowpass_filter=False,
-        high_cut_frequency=210*1e3
-    )
-)
+]
 
 
 # Survey
 survey = em1d.survey.EM1DSurveyTD(source_list)
-
-###############################################
-# Plot the Waveforms
-# ------------------
-#
-#
-
-fig = plt.figure(figsize=(6, 4))
-ax = fig.add_axes([0.1, 0.1, 0.85, 0.8])
-ax.plot(np.r_[-1e-2, 0., 1e-10, 1e-3], np.r_[1., 1., 0., 0.], 'k', lw=2)
-ax.plot(time_input_currents, input_currents, 'b', lw=2)
-ax.plot(time_input_currents_2, input_currents_2, 'r', lw=2)
-ax.set_xlabel("Time (s)")
-ax.set_ylabel("Normalized Current (A)")
-ax.legend(["Step-off", "Triangular", "VTEM"])
-
-
-
 
 
 ###############################################
@@ -135,7 +84,7 @@ n_layer = len(thicknesses) + 1
 sigma = 1e-2
 eta = 0.5
 tau = 0.01
-c = 0.5
+c = 0.75
 chi = 0.
 
 # physical property models
@@ -154,9 +103,9 @@ sigma_complex = ColeCole(frequencies, sigma, eta, tau, c)
 
 fig = plt.figure()
 ax = fig.add_subplot(111)
-ax.semilogx(frequencies, sigma*np.ones(len(frequencies)), "b", lw=2)
-ax.semilogx(frequencies, np.real(sigma_complex), "r", lw=2)
-ax.semilogx(frequencies, np.imag(sigma_complex), "r--", lw=2)
+ax.semilogx(frequencies, sigma*np.ones(len(frequencies)), "b", lw=3)
+ax.semilogx(frequencies, np.real(sigma_complex), "r", lw=3)
+ax.semilogx(frequencies, np.imag(sigma_complex), "r--", lw=3)
 ax.set_xlim(np.min(frequencies), np.max(frequencies))
 ax.set_ylim(0., 1.1*sigma)
 ax.set_xlabel("Frequency (Hz)")
@@ -173,31 +122,41 @@ ax.legend(
 
 
 # Simulate response for static conductivity
-simulation = em1d.simulation.EM1DTMSimulation(
+simulation_1 = em1d.simulation.EM1DTMSimulation(
     survey=survey, thicknesses=thicknesses, sigmaMap=model_mapping,
     chi=chi_model
 )
 
-dpred = simulation.dpred(sigma_model)
+dpred_1 = simulation_1.dpred(sigma_model)
 
-#######################################################################
-# Plotting Results
-# -------------------------------------------------
-#
-#
+# Simulate response for complex conductivity
+simulation_2 = em1d.simulation.EM1DTMSimulation(
+    survey=survey, thicknesses=thicknesses, sigmaMap=model_mapping,
+    eta=eta, tau=tau, c=c, chi=chi_model
+)
+
+dpred_2 = simulation_2.dpred(sigma_model)
 
 
 fig = plt.figure(figsize = (6, 5))
-d = np.reshape(dpred, (len(source_list), len(times))).T
 ax = fig.add_axes([0.1, 0.1, 0.8, 0.85])
-colorlist = ['k', 'b', 'r']
-for ii, k in enumerate(colorlist):
-    ax.loglog(times, np.abs(d[:, ii]), k, lw=2)
+ax.loglog(times, np.abs(dpred_1[0:len(times)]), 'b', lw=3)
+ax.loglog(times, np.abs(dpred_2[0:len(times)]), 'r', lw=3)
+ax.legend(["Non-Chargeable", "Chargeable"])
+ax.set_xlabel("Times (s)")
+ax.set_ylabel("|B| (T)")
+ax.set_title("Magnetic Flux")
 
-ax.legend(["Step-off", "Triangular", "VTEM"])
+fig = plt.figure(figsize = (6, 5))
+ax = fig.add_axes([0.1, 0.1, 0.8, 0.85])
+ax.loglog(times, np.abs(dpred_1[len(times):]), 'b', lw=3)
+ax.loglog(times, np.abs(dpred_2[len(times):]), 'r', lw=3)
+ax.legend(["Non-chargeable", "Chargeable"])
 ax.set_xlabel("Times (s)")
 ax.set_ylabel("|dB/dt| (T/s)")
+ax.set_title("Time-Derivative of Magnetic Flux")
 
+##################################################
